Chapter-1
How to create a blockchain and how it works.
It's completely related to Hashing and how one block is connected o another block. 
the code is completely written in Python using Spyder IDE. 

Note: It's not my creation i have made this when i was learning about blockchains and bitcoins(i.e fintech).